Multiple content updates (#57)
Replaced "Disclosure" with "LinkOut" in example on index page. Removed note about build errors from LinkOut page since the issue is fixed. Overhauled the Markdown syntax page as the info was more general to markdown than it is to the ILF. Added note to the Disclosure page that we should use Details instead, and that the Disclosure page may be going away. Added a Details page. Changed references to running prettier to pnpm format since it's easier to remember. Added a page specifically for Starlight components.